move components to corresponding folders
